Move RpcFile to api codebase
