+++ src/api/services/map/service.ts
@@ -4,7 +4,7 @@ import { defined } from "../../../lib/std/defined";
 import { createSearchProcedure } from "../../common/search";
 import { bufferToLuaCode } from "../../common/parseLuaTableAs";
 import { t } from "../t";
-import { rpcFile } from "../../../lib/rpc/RpcFile";
+import { rpcFile } from "../../common/RpcFile";
 import { access } from "../../middlewares/access";
 import { UserAccessLevel } from "../user/types";
 import { MapRepository } from "./repository";
